+++ HISTORY.md
@@ -3,6 +3,7 @@ unreleased
 
   * Fix check for default `secure` option behavior
   * Fix `maxAge` option preventing cookie deletion
+  * Support `"none"` in `sameSite` option
   * deps: depd@~2.0.0
     - Replace internal `eval` usage with `Function` constructor
     - Use instance methods on `process` to check for listeners